+++ include/mdintg.h
@@ -34,7 +34,4 @@ constexpr unsigned RESPA_FAST = 1; // 2**0, fast group shall be 0.
 constexpr unsigned RESPA_SLOW = 2; // 2**1, slow group shall be 1.
 const TimeScaleConfig& respa_tsconfig();
 void respa_fast_slow(int istep, time_prec dt_ps);
-
-
-void vv_lpiston_npt(int istep, time_prec dt_ps);
 }
