+++ src-acc/lpiston_acc.cpp
@@ -1,3 +1,4 @@
+#include "add.h"
 #include "box.h"
 #include "energy.h"
 #include "mdcalc.h"
@@ -15,6 +16,73 @@
 
 
 namespace tinker {
+
+
+//====================================================================//
+
+
+void ratcom_kevir_acc(double coef, double atomic_vir, double& val)
+{
+   int nmol = rattle_dmol.nmol;
+   auto* molmass = rattle_dmol.molmass;
+   auto* imol = rattle_dmol.imol;
+   auto* kmol = rattle_dmol.kmol;
+
+   double half_coeff = 0.5 * coef;
+
+   double sum = 0;
+   #pragma acc parallel loop independent async\
+               copy(sum) reduction(+:sum)\
+               deviceptr(molmass,imol,kmol,ratcom_px,ratcom_py,ratcom_pz,\
+                  ratcom_x,ratcom_y,ratcom_z,ratcom_gx,ratcom_gy,ratcom_gz,\
+                  xpos,ypos,zpos,gx,gy,gz)
+   for (int im = 0; im < nmol; ++im) {
+      auto px = ratcom_px[im];
+      auto py = ratcom_py[im];
+      auto pz = ratcom_pz[im];
+      auto mmassinv = 1.0 / molmass[im];
+      auto ke_x2 = (px * px + py * py + pz * pz) * mmassinv;
+
+      double igx, igy, igz, mgx, mgy, mgz;
+      double mvir = 0.0;
+#if TINKER_DETERMINISTIC_FORCE
+      mgx = to_flt_acc<double>(ratcom_gx[im]);
+      mgy = to_flt_acc<double>(ratcom_gy[im]);
+      mgz = to_flt_acc<double>(ratcom_gz[im]);
+#else
+      mgx = ratcom_gx[im];
+      mgy = ratcom_gy[im];
+      mgz = ratcom_gz[im];
+#endif
+      mvir += mgx * ratcom_x[im] + mgy * ratcom_y[im] + mgz * ratcom_z[im];
+
+      int start = imol[im][0];
+      int stop = imol[im][1];
+      #pragma acc loop seq
+      for (int i = start; i < stop; ++i) {
+         int k = kmol[i];
+#if TINKER_DETERMINISTIC_FORCE
+         igx = to_flt_acc<double>(gx[k]);
+         igy = to_flt_acc<double>(gy[k]);
+         igz = to_flt_acc<double>(gz[k]);
+#else
+         igx = gx[k];
+         igy = gy[k];
+         igz = gz[k];
+#endif
+         mvir -= (igx * xpos[k] + igy * ypos[k] + igz * zpos[k]);
+      }
+
+      sum += half_coeff * ke_x2 - mvir;
+   }
+   #pragma acc wait
+
+   // sum = coef * kinetic + molecular_vir - atomic_vir
+   // return: sum + atomic_vir
+   val = sum + atomic_vir;
+}
+
+
 namespace {
 void vvlp_hc_pt(time_prec dt, double R, int)
 {
